TradingAgents/requirements.txt
# ============================================================
# Modified: See CHANGELOG.md for complete modification history
# Last Updated: 2025-10-25
# Modified By: jimyungkoh<aqaqeqeq0511@gmail.com>
# ============================================================

typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
pre-commit
google-genai
fastapi
uvicorn
python-dotenv
pytest
